added useParentForScrolling option
Added an option "useParentForScrolling" to scrolling calculations to be
done with the parent of the widget, for situations where you want to use
a wrapper div. (See
http://stackoverflow.com/questions/2330773/jquery-ui-selectable-plugin-make-scroll-bar-not-selectable-when-div-overflows
for an example. This is something you might want to do if you don't want
clicks on the scrollbar to be treated as selections in the selectable.)